mise a jour

+++ database/migrations/2024_02_16_000004_create_reviews_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id('idReview');
             $table->unsignedBigInteger('user_id');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->integer('rating');
             $table->text('comment')->nullable();
             $table->boolean('is_approved')->default(false);
